+++ samples/XCT.Sample/App.xaml.cs
@@ -1,5 +1,6 @@
 ﻿using System.Globalization;
 using Xamarin.CommunityToolkit.Helpers;
+using Xamarin.CommunityToolkit.PlatformConfiguration.Multiplatform;
 using Xamarin.CommunityToolkit.Sample.Pages;
 using Xamarin.CommunityToolkit.Sample.Resx;
 using Xamarin.Forms.PlatformConfiguration;
@@ -19,6 +20,8 @@ namespace Xamarin.CommunityToolkit.Sample
 
 			InitializeComponent();
 			MainPage = new BaseNavigationPage(new WelcomePage());
+
+			var tmp = DeviceCulture.GetPreferredCultures();
 		}
 	}
 }
